refactor: user 엔티티 트랜잭션 추가 및 파일명 변경, 회원가입 로직에 온도 추가 #1

+++ src/main/java/unilearn/unilearn/user/entity/Temperature.java
@@ -18,5 +18,5 @@ public class Temperature {
     @JoinColumn(name = "user_id")
     private User user;
     @ColumnDefault("36.5")
-    private float temperature;
+    private double temperature;
 }
